requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anyio==4.7.0
attrs==24.2.0
awacs==2.4.1
aws-sam-translator==1.94.0
billiard==4.2.1
blinker==1.9.0
boto3==1.35.78
botocore==1.35.78
bracex==2.5.post1
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
cfn-flip==1.3.0
cfn-lint==1.22.0
charset-normalizer==3.4.0
cli_exit_tools==1.2.7
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
coloredlogs==15.0
cryptography==44.0.0
distlib==0.3.9
distro==1.9.0
docker==7.1.0
filelock==3.16.1
Flask==3.1.0
Flask-Mail==0.10.0
Flask-SQLAlchemy==3.1.1
formic2==1.0.3
gitdb==4.0.11
GitPython==3.1.43
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
humanfriendly==10.0
idna==3.10
igittigitt==2.1.5
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.8.2
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kombu==5.4.2
lark==1.2.2
lib-detect-testenv==2.0.8
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
openai==1.57.2
packaging==24.2
pathspec==0.12.1
pipenv==2022.1.8
platformdirs==4.3.6
prompt_toolkit==3.0.48
pycparser==2.22
pydantic==2.10.3
pydantic_core==2.27.1
pyhcl==0.4.5
pyOpenSSL==24.3.0
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-hcl2==5.1.1
pywin32==308
PyYAML==6.0.2
redis==5.2.1
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
rpds-py==0.22.3
runway==2.7.0
s3transfer==0.10.4
Send2Trash==1.8.3
setuptools==75.6.0
six==1.17.0
smmap==5.0.1
sniffio==1.3.1
SQLAlchemy==2.0.36
sympy==1.13.3
tomli==2.2.1
tqdm==4.67.1
troposphere==4.8.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
utaw==0.2.0
vine==5.1.0
virtualenv==20.28.0
virtualenv-clone==0.5.7
wcmatch==10.0
wcwidth==0.2.13
Werkzeug==3.1.3
yamllint==1.35.1
